Obvious fix for #9550 (m_totalSize is already an unsigned long)
--- kdirsize.h  2000/05/17 01:20:30     1.1
+++ kdirsize.h  2000/09/05 13:19:19
   /**
    * @return the size we found
    */
-  int totalSize() { return m_totalSize; }
+  unsigned long totalSize() { return m_totalSize; }

--- kpropsdlg.cpp       2000/08/26 12:29:49     1.145
+++ kpropsdlg.cpp       2000/09/05 13:19:37
@@ -777,7 +777,7 @@ void KFilePropsPlugin::slotDirSizeFinish
     job->showErrorDialog( properties->dialog() );
   else
   {
-    int totalSize = static_cast<KDirSize*>(job)->totalSize();
+    unsigned long totalSize = static_cast<KDirSize*>(job)->totalSize();
     m_sizeLabel->setText( QString::fromLatin1("%1 (%2)").arg(KIO::convertSize(totalSize)).arg(totalSize) );

svn path=/trunk/kdelibs/; revision=62752

+++ kfile/kdirsize.h
@@ -37,7 +37,7 @@ public:
   /**
    * @return the size we found
    */
-  int totalSize() { return m_totalSize; }
+  unsigned long totalSize() { return m_totalSize; }
 
   /**
    * Asynchronous method. Connect to the result signal.
